Develop

+++ RetailStore/Requests/CustomerManagement/DeleteCustomerCommand.cs
@@ -1,33 +1,39 @@
 ﻿using MediatR;
+using RetailStore.Constants;
 using RetailStore.Model;
-using RetailStore.Persistence;
 using RetailStore.Repository;
-using RetailStore.Requests.OrderManagement;
+using Twilio.TwiML.Messaging;
 
-namespace RetailStore.Features.CustomerManagement;
+namespace RetailStore.Requests.CustomerManagement;
 
 /// <summary>
 /// Command to Delete Customer by Id
 /// </summary>
-public class DeleteCustomerCommand : IRequest<Customer>
+public class DeleteCustomerCommand : IRequest<string>
 {
+    /// <summary>
+    /// Gets and sets Id
+    /// </summary>
     public int CustomerId { get; set; }
 }
 
 /// <summary>
 /// Handler for Delete Customer by Id command
 /// </summary>
-public class DeleteCustomerCommandHandler : IRequestHandler<DeleteCustomerCommand, Customer>
+public class DeleteCustomerCommandHandler : IRequestHandler<DeleteCustomerCommand, string>
 {
     private readonly IRepository<Customer> _customerRepository;
+    private readonly ILogger<DeleteCustomerCommandHandler> _logger;
 
     /// <summary>
     /// Injects IRepository class
     /// </summary>
     /// <param name="customerRepository"></param>
-    public DeleteCustomerCommandHandler(IRepository<Customer> customerRepository)
+    /// <param name="logger">The logger instance.</param>
+    public DeleteCustomerCommandHandler(IRepository<Customer> customerRepository, ILogger<DeleteCustomerCommandHandler> logger)
     {
         _customerRepository = customerRepository;
+        _logger = logger;
     }
 
     /// <summary>
@@ -35,18 +41,20 @@ public class DeleteCustomerCommandHandler : IRequestHandler<DeleteCustomerComman
     /// </summary>
     /// <param name="command"></param>
     /// <param name="cancellationToken"></param>
-    /// <returns>Customer</returns>
-    /// <exception cref="KeyNotFoundException"></exception>
-    public async Task<Customer> Handle(DeleteCustomerCommand command, CancellationToken cancellationToken)
+    /// <returns>Response message</returns>
+    /// <exception cref="ValidationException"></exception>
+    public async Task<string> Handle(DeleteCustomerCommand command, CancellationToken cancellationToken)
     {
-        var deletedCustomer = await _customerRepository.Delete(command.CustomerId);
+        var customer = await _customerRepository.GetById(command.CustomerId);
 
-        if (deletedCustomer == null)
+        if (customer == null)
         {
-            throw new KeyNotFoundException();
+            return string.Format(ValidationMessage.CustomerDoesNotExist, command.CustomerId);
         }
 
-        return deletedCustomer;
+        var deletedCustomer = await _customerRepository.Delete(command.CustomerId);
 
+        _logger.LogInformation(LogMessage.DeleteItem, deletedCustomer.Id);
+        return string.Format(ValidationMessage.CustomerDeletedSuccessfully, command.CustomerId);
     }
 }
